Refactor webhonc message handlers
